MYFACES-4458 for 2.3 next (#357)
* MYFACES-4458 for 2.3 next

* 2.13.3.Final
