+++ extensions/quarkus/runtime/pom.xml
@@ -56,8 +56,18 @@
 
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-undertow-websockets</artifactId>
+            <artifactId>quarkus-arc</artifactId>
             <version>${quarkus.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-core</artifactId>
+            <version>${quarkus.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-undertow-websockets</artifactId>
+            <version>2.7.6.Final</version>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
@@ -138,10 +148,11 @@
         <plugins>
             <plugin>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-bootstrap-maven-plugin</artifactId>
+                <artifactId>quarkus-extension-maven-plugin</artifactId>
                 <version>${quarkus.version}</version>
                 <executions>
                     <execution>
+                        <phase>compile</phase>
                         <goals>
                             <goal>extension-descriptor</goal>
                         </goals>
@@ -150,7 +161,8 @@
                         </configuration>
                     </execution>
                 </executions>
-            </plugin>  
-        </plugins>        
+            </plugin>
+        </plugins>
+
     </build>
 </project>
